task completed

+++ employeeapp/views.py
@@ -1,15 +1,12 @@
 '''Employees views'''
-import re
-from urllib import response
 from rest_framework.generics import GenericAPIView
 from rest_framework.response import Response
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 from uritemplate import partial
-from .serializers import EmployeeSerializer, Employee, EmployeeUpdateSerializer
+from .serializers import EmployeeSerializer, Employee
 from rest_framework import status
-from rest_framework.parsers import MultiPartParser
-from django.shortcuts import  get_object_or_404
+from django.shortcuts import get_object_or_404
 
 employee_get_update_response = {
     "200": EmployeeSerializer,
@@ -22,7 +19,6 @@ employee_create_response = {
 }
 
 class EmployeeView(GenericAPIView):
-    parser_classes =[MultiPartParser]
     serializer_class = EmployeeSerializer
     @swagger_auto_schema(tags=["Employee"], operation_description="Emloyee List", operation_summary="Employee List", responses=employee_get_update_response)
     def get(self, request):
@@ -46,10 +42,9 @@ class EmployeeView(GenericAPIView):
         return Response(response,http_status)
 
 class EmployeeDetailView(GenericAPIView):
-    # parser_classes = [MultiPartParser]
     serializer_class = EmployeeSerializer
     def get_employee(self, id):
-        employee = Employee.objects.filter(id=id).first()
+        employee = get_object_or_404(Employee, pk=id)
         return employee
     @swagger_auto_schema(tags=["Employee"], operation_description="Emloyee Get", operation_summary="Employee Get", responses=employee_get_update_response)
     def get(self, request, id):
@@ -77,29 +72,25 @@ class EmployeeDetailView(GenericAPIView):
         response = {}
         http_status = None
         employee = self.get_employee(id)
-        serialized_data = EmployeeUpdateSerializer(employee, data=request.data, partial=True)
+        serialized_data = self.serializer_class(employee, data=request.data, partial=True)
         serialized_data.is_valid(raise_exception=True)
         serialized_data.save()
         response["message"] = "Data updated"
         response["data"] = serialized_data.data
         http_status = status.HTTP_200_OK
         return Response(response, http_status)
-    @swagger_auto_schema(tags=["Employee"], operation_description="Emloyee Delete", operation_summary="Employee Delete")
+    @swagger_auto_schema(tags=["Employee"], operation_description="Employee Delete", operation_summary="Employee Delete")
     def delete(self, request, id):
         response = {}
         http_status = None
         employee = self.get_employee(id)
-        employee.delete()
-        response["message"] = "data deleted"
-        http_status = status.HTTP_200_OK
+        if employee:
+            employee.delete()
+            response["message"] = "data deleted"
+            http_status = status.HTTP_200_OK
+        else:
+            response["message"] = "data not found"
+            http_status = status.HTTP_404_NOT_FOUND
         return Response(response, http_status)
-
-
-
-        
-
-
-
-
